requirements_django_app.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
better-abc==0.0.3
blinker==1.8.2
boto3==1.34.122
botocore==1.34.122
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
cryptography==42.0.8
datasets==2.20.0
debugpy==1.6.7
decorator==5.1.1
diffusers==0.29.2
dill==0.3.8
Django==4.2.13
django-allauth==0.63.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
ecdsa==0.19.0
entrypoints==0.4
evaluate==0.4.2
executing==2.0.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.12.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter-client==7.3.4
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mongoengine==0.28.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
onnx==1.16.1
onnxruntime-gpu==1.18.1
opencv-python==4.10.0.82
optimum==1.21.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
python-dateutil==2.9.0
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2024.5.15
requests==2.32.3
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.2
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
